v2.9.13

+++ snort/tools/chocolateyuninstall.ps1
@@ -1,3 +1,4 @@
-﻿If (Test-Path c:\Snort){
+﻿$ErrorActionPreference = 'Stop'
+If (Test-Path c:\Snort){
    Remove-Item -path c:\Snort -recurse
 }
